Add example with vertical bars in BarSeriesExamples.cs
In the past, there was ColumnSeries, which could have the categories be placed on the X axis. The BarSeries needs them on the Y axis. The new example show how to do this, but use keys to have to categories on bottom axis in the plot anyway.

Update CONTRIBUTORS

Update CHANGELOG.md